% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiMatrixMC.R
\name{ChiMatrixMC}
\alias{ChiMatrixMC}
\title{An empirical chi matrix}
\usage{
ChiMatrixMC(Dat, quan = NULL)
}
\arguments{
\item{Dat}{An \eqn{n\times d} data matrix from the inverted Pareto distribution.}

\item{quan}{Numeric; a lower threshold quantile}
}
\value{
A \eqn{d\times d} strict upper triangular matrix of empirical chi matrix.
}
\description{
The function \code{ChiMatrixMC} takes a inverted-Pareto data matrix and returns an empirical chi matrix regardless of rank transformations.
}
\examples{
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# Pareto random samples
Dat_P=ParetoSim(n = 5000, XVS = XVS) # Pareto scale
ChiMatrixMC(1/Dat_P)
}
