% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiMtxMC.R
\name{ChiMtxMC}
\alias{ChiMtxMC}
\title{Pairwise tail dependence matrix via Monte Carlo simulations}
\usage{
ChiMtxMC(Dat, quan = NULL)
}
\arguments{
\item{Dat}{An \eqn{n\times d} data matrix from multivariate inverted-Pareto distribution.
If the original data are used, then the data transformation is performed to obtain pseudo-observations for the inverted-Pareto distribution
See \code{ParetoTransRank()}.}

\item{quan}{A numeric quantile, set a lower enough quantile \eqn{u_{quan}\in(0,1)} to convert data into pseudo-observations for the inverted-Pareto distribution.
Default is \code{quan=NULL}.}
}
\value{
A \eqn{d\times d} strict upper triangular matrix of empirical pairwise chi-matrix.
}
\description{
\code{ChiMtxMC} takes an inverted-Pareto data matrix as input and computes an empirical pairwise tail dependence measure via Monte Carlo simulations.
The resulting matrix is upper strict and computes the tail dependence coefficient between variables numerically.
Each element in the chi-matrix consists of the (lower) tail dependence coefficient \eqn{\chi_{L,e}=R(1,1)} for \eqn{e\in T_i}, \eqn{i=1,\ldots,d-1}.
}
\examples{
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# Pareto random samples
Dat_P=ParetoSim(n = 2000, XVS = XVS) # Pareto scale
# Pairwise chi-matrix
ChiMtxMC(Dat_P)
}
\seealso{
\code{\link[=ParetoTransRank]{ParetoTransRank()}}
}
