% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiExpFamily.R
\name{CondExp}
\alias{CondExp}
\title{Bivariate tail copula with a fixed argument}
\usage{
CondExp(x1, x2, par, family)
}
\arguments{
\item{x1}{A numeric vector of the first argument for the bivariate tail copula.}

\item{x2}{A numeric vector of the second argument for the bivariate tail copula.}

\item{par}{A numeric value of parameter values.}

\item{family}{A numeric vector indicating the list of bivariate tail copulas.
The available tail copula families are:
\itemize{
\item 1=Husler-Reiss
\item 2=Negative logistic
\item 3=Logistic
\item 4=Dirichlet
}}
}
\value{
A numeric vector of the bivariate tail copula with a fixed argument.
}
\description{
\code{CondExp()} evaluates the univariate distribution of the bivariate tail copula
given one of its arguments is fixed, \eqn{\Lambda_{1|2}(u_1=x_1|x_2;\theta)}.
}
