% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condfit.FirstTree.R
\name{Condfit.FirstTree}
\alias{Condfit.FirstTree}
\title{Redefine pseudo-observations for the conditional tail copula above the second level tree}
\usage{
Condfit.FirstTree(condSet, data, VineTree)
}
\arguments{
\item{condSet}{A numeric vector of the conditioning set where the cardinality is greater than 2.}

\item{data}{A \eqn{n\times d} inverted-Pareto sample data set.}

\item{VineTree}{A list of selected vine structure and fitted vine models.}
}
\value{
A list of pseudo-observations for the conditional tail copula:
\itemize{
\item Pseudo1
\item Pseudo2
}
}
\description{
\code{Condfit.FirstTree()} redefine pseudo-observations for the conditional tail copula above the second level tree
where the cardinality of the conditioning set is greater than 2.
}
