% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condfit.SecondTree.R
\name{Condfit.SecondTree}
\alias{Condfit.SecondTree}
\title{Define pseudo-observations of the conditional pair-copulas for the second tree}
\usage{
Condfit.SecondTree(tree, data, MST, VineTree)
}
\arguments{
\item{tree}{Numeric; a tree level.}

\item{data}{A \eqn{n\times d} inverted-Pareto sample data set.}

\item{MST}{A list of maximum spanning trees.}

\item{VineTree}{A list of selected vine structures and fitted X-vine models.}
}
\value{
A list of pseudo-observations such that two variables are less than 1 where two variables are nodes in \eqn{T_2}.
These pseudo-data will be used for estimating pair-copulas in the next tree \eqn{T_3}.
}
\description{
\code{Condfit.SecondTree()} defines pseudo-observations for the conditional pair-copulas with arguments properly conditioned.
For each edge, the conditioning set\code{condSet} is passed to the function \code{Condfit.FirstTree()} where we condition two variables in the conditioning set being less than 1.
Note that the previously defined pseudo-observations are conditioned on one variable being less than 1, which are used to estimate pair-copulas in \eqn{T_2} but not in \eqn{T_3}.
}
