% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condfit.SubTree.R
\name{Condfit.SubTree}
\alias{Condfit.SubTree}
\title{Redefine pseudo-observations for the conditional pair-copula above the second level tree}
\usage{
Condfit.SubTree(tree, MST, data, VineTree)
}
\arguments{
\item{tree}{Numeric; a tree level.}

\item{MST}{A list of maximum spanning trees.}

\item{data}{A \eqn{n\times d} inverted-Pareto sample data set.}

\item{VineTree}{A list of selected vine structures and fitted X-vine models.}
}
\value{
A list of redefined pseudo-observations for the conditional pair-copulas with all variables in the conditioning set being less than 1.
}
\description{
\code{Condfit.SubTree()} defines pseudo-observations for the conditional pair-copulas with arguments properly conditioned.
For each edge, the conditioning set\code{condSet} is passed to the function \code{pseudo.ite()} where we condition all variables in the conditioning set being less than 1.
Note that the previously defined pseudo-observations are conditioned on one less variable being less than 1, which are used to estimate pair-copulas in \eqn{T_{i-1}} but not in \eqn{T_i}.
}
