% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiTCfamily.R
\name{CondTC}
\alias{CondTC}
\title{Bivariate tail copula with the second argument being conditioned}
\usage{
CondTC(x1, x2, par, family)
}
\arguments{
\item{x1}{A numeric vector of the first argument for the bivariate tail copula.}

\item{x2}{A numeric vector of the second argument for the bivariate tail copula.}

\item{par}{A numeric value of parameter values.}

\item{family}{A numeric vector indicating the list of bivariate tail copulas.
The available tail copula families are:
\itemize{
\item 1=Husler-Reiss
\item 2=Negative logistic
\item 3=Logistic
\item 4=Dirichlet
}}
}
\value{
A numeric vector of the bivariate tail copula with a fixed argument.
}
\description{
\code{CondTC()} evaluates the univariate distribution of the bivariate tail copula
given one of its arguments is less than 1, \eqn{\Lambda_{1|2}(u_1=x_1|x_2;\theta)}
for \eqn{0< x_2 <1}.
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.
}
