% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependMtx.R
\name{DependMtx}
\alias{DependMtx}
\title{Dependence measure matrix for X-vine models}
\usage{
DependMtx(XVS)
}
\arguments{
\item{XVS}{A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:\code{\link[=XVineSpec]{XVineSpec()}}.}
}
\value{
A \eqn{d\times d} upper triangular matrix of dependence measures
using the parameter relationships.
}
\description{
\code{DependMtx} provides a summary matrix of bivariate dependence measures for X-vine models, using parametric relationships between
parameters associated with (tail) copula models and (tail) dependence measures.
The first row includes the (lower) tail dependence coefficient \eqn{\chi_{L,e}=R(1,1)} for \eqn{e\in T_1}.
Possible tail copula families include:
Husler-Reiss, Negative logistic, logistic, and Dirichlet models.
For explicit formulas, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023).
and the remaining rows include the measure of Kendall's tau, \eqn{\tau_{e}}, for \eqn{e\in T_i}, \eqn{i=2,\ldots,d-1}.
Any copula family with a single parameter can be considered. For more details, refer to Czado (2019, Section 3.5).
}
\examples{
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine speicification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# Dependence matrix for X-vine models
DependMtx(XVS)
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.
Czado, C. (2019). Analyzing dependent data with vine copulas. Lecture Notes in Statistics, Springer, 222.
}
