% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.FirstTree.R
\name{fit.FirstTree}
\alias{fit.FirstTree}
\title{Fit the first maximum spanning tree to multivariate inverted-Pareto data}
\usage{
fit.FirstTree(MST, data.invPa, tcfamset, si, selectioncrit, weights = NA)
}
\arguments{
\item{MST}{A list of maximum spanning trees.}

\item{data.invPa}{A \eqn{n\times d} data matrix from multivariate inverse-Pareto distribution.}

\item{tcfamset}{Numeric vector; specifies the class of bivariate tail copula models that users consider.}

\item{si}{Numeric; a tuning parameter for mBIC (\eqn{\si_0=0.9}; default).}

\item{selectioncrit}{Character string, indicating the selection criteria for the class of bivariate tail copula models.}

\item{weights}{Logical; whether weights should be assigned to observations when missing values exist.}
}
\value{
a nested list object:
\itemize{
\item the maximum spanning tree with a node set and edge set.
\item the fitted tail copula model with selected classes, parameter estimates, log-likelihood values, AIC, and pseudo-observations.
}
}
\description{
\code{fit.FirstTree()} fits the first maximum spanning tree to multivariate inverted-Pareto data set.
The function takes a pair of variables for each edge and variables are passed to the function 'tcSelect'.
After selecting the family of bivariate tail copula models, store the output of the fitted model in the list of MST
, including parameter estimates, model classes, fitted models, and pseudo-observations.
}
