% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiExpFamily.R
\name{InvExp}
\alias{InvExp}
\title{Quantile function of a univariate tail copula}
\usage{
InvExp(w2, u1, par, family)
}
\arguments{
\item{w2}{A numeric vector of the first argument for the bivariate tail copula.}

\item{u1}{A numeric vector of the second argument conditioned on \eqn{u_1 < 1}
for the bivariate tail copula.}

\item{par}{A numeric of parameter values.}

\item{family}{A numeric vector indicating the list of tail copula families.}
}
\value{
A numeric vector of quantile values.
}
\description{
Conditioning on the second argument of the bivariate tail copula being less than 1,
\eqn{x_1 < 1}, the function \code{InvExp()} computes the quantiles of the univariate tail copula,
\eqn{r^{-1}_{2|1}(w_2|u_1=x_1;\theta)}.
The available tail copula families are:
\itemize{
\item 1=Husler-Reiss
\item 2=Negative logistic
\item 3=Logistic
\item 4=Dirichlet
}
}
