% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiLoglik.R
\name{LL.BiTC}
\alias{LL.BiTC}
\title{Log-likelihood function for bivariate tail copula densities on product space}
\usage{
LL.BiTC(par, x, family)
}
\arguments{
\item{par}{Numeric; a parameter for bivariate tail copulas}

\item{x}{A \eqn{N\times 2} data matrix from a bivariate inverted-Pareto distribution
where the second column is less than 1}

\item{family}{An integer; indicates the type of bivariate tail copula families.
Possible tail copula families include:
1=Husler-Reiss
2=Negative Logistic
3=Logistic
4=Dirichlet}
}
\value{
Numeric; a negative log-likelihood value
}
\description{
\code{LL.BiTC} calculates a negative log-likelihood of a bivariate tail copula density \eqn{r} on product space with uniform margins.
As the bivariate tail copula \eqn{r(x1,x2)} with \eqn{x2<1} corresponds to the bivariate inverted-Pareto density,
the normalizing constant can be ignored because it is 1.
This function is passed to the function \code{mleBiTC()} for finding ML estimates.
}
