% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBICplot.R
\name{mBICplot}
\alias{mBICplot}
\title{Sensitive analysis for the mBIC-optimal truncation level across low quantiles.}
\usage{
mBICplot(
  data,
  Rank = TRUE,
  quan,
  famset_tc = c(1:4),
  famset_pc = c(0, 1, 3, 4, 5, 6, 13, 14, 16),
  selectioncrit = "AIC",
  trunclevel = FALSE,
  treecritT1 = "chi",
  treecritT2 = "tau",
  effsampsize = 10,
  tau_threshold = 0.05,
  si = 0.9
)
}
\arguments{
\item{data}{A \eqn{n\times d} data matrix that is passed into the function \code{XVineModelFit_lightversion()}
performs the rank transformation.}

\item{Rank}{Logical; whether rank transformation is performed or not (\code{Rank=T}; default).}

\item{quan}{A numeric vector; a sequence of lower quantiles in ascending order.}

\item{famset_tc}{Numeric vector; the class of bivariate exponent measures.}

\item{famset_pc}{Numeric vector; the class of bivariate pair-copula models.}

\item{selectioncrit}{Character string; indicates the selection criteria for the class of bivariate exponent measures or pair-copulas.}

\item{trunclevel}{Numeric; indicates the specified truncation level (\code{trunclevel=NULL}; default).}

\item{treecritT1}{A character string, indicating the tree criterion for the first tree.}

\item{treecritT2}{A character string, indicating the tree criterion for subsequent trees.}

\item{effsampsize}{Numeric; the specified effective sample size for the independence copula (\eqn{n_{D_e}}<10; default).}

\item{tau_threshold}{Numeric; the specified Kendall's tau value for the independence copula (\eqn{\hat{\tau}_e}<0.05; default).}

\item{si}{Numeric; a tuning parameter for mBIC (\eqn{\si_0=0.9}; default).}
}
\value{
A \code{mBICvsQuan} object
}
\description{
\code{mBICplot()} plots the mBIC-optimal truncation levels across low quantiles
, which can be used for sensitive analysis in data application.
}
\examples{
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
Dat_P=ParetoSim(n = 2000, XVS = XVS) # Pareto scale
quan=seq(0.05,0.15,by=0.01)
mBICplot(Dat_P, Rank=TRUE, quan
        , famset_tc=c(1:4), famset_pc=c(0,1,3,4,5,6,13,14,16)
        , selectioncrit="AIC",trunclevel=FALSE
        , treecritT1="chi",treecritT2="tau"
        , effsampsize=10, tau_threshold=0.05, si=0.9)
}
