% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2DepMtx.R
\name{ML2DepMtx}
\alias{ML2DepMtx}
\title{Converting ML estimates to dependence measures}
\usage{
ML2DepMtx(FamMtx, ParMtx)
}
\arguments{
\item{FamMtx}{A \eqn{d\times d} upper triangular family matrix indicating the family of bivariate (tail) copula models}

\item{ParMtx}{A \eqn{d\times d} upper triangular parameter matrix including ML estimates for each edge}
}
\value{
A \eqn{d\times d} upper triangular matrix of dependence measures
}
\description{
\code{ML2DepMtx} takes a family matrix and parameter matrix and
converts ML estimates for X-vine models to dependence measures via parametric relationships between
parameters associated with (tail) copula models and (tail) dependence measures.
The first row includes the (lower) tail dependence coefficient \eqn{\chi_{L,e}=R(1,1)} for \eqn{e\in T_1}.
Possible tail copula families include:
Husler-Reiss, Negative logistic, logistic, and Dirichlet models.
For explicit formulas, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023).
and the remaining rows include the measure of Kendall's tau, \eqn{\tau_{e}}, for \eqn{e\in T_i}, \eqn{i=2,\ldots,d-1}.
Any copula family with a single parameter can be considered. For more details, refer to Czado (2019, Section 3.5).
}
\examples{
ParamM <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamilyM <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
ML2DepMtx(FamMtx=FamilyM,ParMtx=ParamM)               
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.
Czado, C. (2019). Analyzing dependent data with vine copulas. Lecture Notes in Statistics, Springer, 222.
}
