% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mleBiTC}
\alias{mleBiTC}
\title{Maximum likelihood estimates for bivariate (tail) copula densities}
\usage{
mleBiTC(ft, family, data, range)
}
\arguments{
\item{ft}{A log-likelihood function (see: \code{BiLoglik()})}

\item{family}{An integer; indicates the type of bivariate (tail) copula densities.
Possible tail copula families include:
1=Husler-Reiss
2=Negative Logistic
3=Logistic
4=Dirichlet
Possible copula families include:
0=Independence
1=Gaussian
3=Clayton
4=Gumbel
5=Frank
6=Joe
13=Survival Clayton
14=Survival Gumbel
16=Survival Joe}

\item{data}{An \eqn{N\times 2} data matrix from the bivariate inverted Pareto samples with the second column being less than 1}

\item{range}{A numeric vector; indicates the range of parameters}
}
\value{
A numeric; ML estimate(s)
}
\description{
\code{mleBiTC} returns ML estimates for bivariate tail copula densities and
\code{mleBiCop} returns ML estimates for bivariate copula densities.
}
