% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoTransRank.R
\name{ParetoTransRank}
\alias{ParetoTransRank}
\title{Transform to multivariate Pareto samples via rank transformation}
\usage{
ParetoTransRank(data, u_quan, scaleType = "U")
}
\arguments{
\item{data}{An \eqn{n\times d} data matrix}

\item{u_quan}{A numeric quantile, set a lower enough quantile \eqn{u_{quan}\in(0,1)}}

\item{scaleType}{A character specifying the type of scales for generated samples (the default is uniform scale: scaleType="U").
Other possible scales:
\itemize{
\item scaleType="P": Pareto scale
\item scaleType="E": Exponential scale
}}
}
\value{
An \eqn{N\times d} pseudo-observations for the inverted Pareto distribution (default)
}
\description{
\code{ParetoTransRank} transforms the original data to multivariate Pareto samples with suitable margins (the default: Pareto margins)
via rank transformation.
}
\details{
Let \eqn{\boldsymbol(X)_{i}, i=1,\ldots,n} be independent samples from the unkown distribution \eqn{F}.
Transform to the uniform margin via rank transformation \eqn{\widehat{U}_{i,j}=1-(rank_{i,j}-0.5)/n}, where
\eqn{rank_{i,j}=\sum_{s=1}^{n}\mathbb{I}(X_{s,j}\le X_{i,j})} is the maximal rank of \eqn{X_{i,j}} among \eqn{X_{1,j},\ldots,X_{n,j}}.
The scaled points \eqn{(n/k)\widehat{U}_i} for \eqn{i=1,\ldots,n} such that \eqn{\min\widehat{U}_i<k/n} are pseudo-observations
from the inverted multivariate Pareto distribution. We set \eqn{k\in\{1,\ldots,n\}}
such that \eqn{k} is large and \eqn{k/n} is small.
If we take the reciprocal of the rescaled samples, then we have Pareto scales.
If we take the negative log of the rescaled samples, then we have exponential scales.
}
\examples{
StrMtx <- matrix(c(1, 1, 2, 2, 4,
             0, 2, 1, 3, 2,
             0, 0, 3, 1, 3,
             0, 0, 0, 4, 1,
             0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                  0, 0, 2, 2.5, 0.7,
                  0, 0, 0, 0.4, -0.3,
                  0, 0, 0, 0, 0.1,
                  0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                0, 0, 3, 4, 1,
                0, 0, 0, 3, 1,
                0, 0, 0, 0, 1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
## X-Vine specification (Vine structure, Bivariate parametric families, Parameters)
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Multivariate Pareto samples
Dat_P=ParetoSim(n = 5000, XVS = XVS) # Pareto scale
##  Transform to pseudo-observations for inverted Pareto distribution
Dat_U=ParetoTransRank(data = Dat_P, u_quan = 0.2, scaleType = "U")
Dat_U
}
