% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcFamSel.R
\name{pcFamSel}
\alias{pcFamSel}
\title{Selecting the family of pair-copula models for each edge and fit its model.}
\usage{
pcFamSel(
  u1,
  u2,
  familyset,
  si = 0.9,
  tree,
  selectioncrit = "AIC",
  effsampsize = 10,
  tau_threshold = 0.05,
  se = FALSE
)
}
\arguments{
\item{u1}{A numeric vector, indicating the first argument for the conditional pair-copulas for each edge in each tree.}

\item{u2}{A numeric vector, indicating the second argument for the conditional pair-copulas for each edge in each tree.}

\item{familyset}{A numeric vector, indicating the list of bivariate pair-copulas.}

\item{si}{Numeric; a tuning parameter for mBIC (\eqn{\si_0=0.9}; default).}

\item{tree}{A numeric value indicating the tree level.}

\item{selectioncrit}{Character string, indicating the selection criteria for bivariate pair copulas.}

\item{effsampsize}{Numeric; the specified effective sample size for the independence copula (\eqn{n_{D_e}}<10; default).}

\item{tau_threshold}{Numeric; the specified Kendall's tau value for the independence copula (\eqn{\hat{\tau}_e}<0.05; default).}

\item{se}{Logical; whether standard errors for ML estimators are reported.}
}
\value{
A list object consisting of the class type, parameter estimates, log-likelihood value, AIC, and pseudo-observations for selected bivariate pair-copulas.
}
\description{
Given the family of pair-copula models users consider, \code{pcFamSel()} selects
the family of pair-copula models with the lowest AIC for a specific edge, and returns
parameter estimates, likelihood value, AIC, and pseud-observations used for the next tree.
}
