% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlights_adj.R
\name{plotFlights_adj}
\alias{plotFlights_adj}
\title{Create figures using flight delays data}
\usage{
plotFlights_adj(
  airportIndices = NULL,
  airports_sel = NULL,
  connections_sel = NULL,
  graph = NULL,
  edgeColors,
  plotAirports = TRUE,
  plotConnections = TRUE,
  returnGGPlot = FALSE,
  useAirportNFlights = FALSE,
  useConnectionNFlights = FALSE,
  minNFlights = 0,
  map = "state",
  vertexColors = NULL,
  vertexShapes = NULL,
  xyRatio = NULL,
  clipMap = FALSE,
  useLatex = FALSE,
  edgeAlpha = 0.4
)
}
\arguments{
\item{airportIndices}{Character string; the indices of the airports}

\item{airports_sel}{the set of airports}

\item{connections_sel}{A three column data frame by \code{flightCountMatrixToConnectionList()}}

\item{graph}{An optional \code{igraph::graph} object}

\item{edgeColors}{Vector or symmetric matrix (character or numeric) used as colors for edges/connections}

\item{plotAirports}{Logical; whether to plot the specified airports}

\item{plotConnections}{Logical; whether to plot the specified connections}

\item{returnGGPlot}{a \code{ggplot2::ggplot} object is return if \code{TRUE}}

\item{useAirportNFlights}{Logical; whether to change the size of the circles indicating airports}

\item{useConnectionNFlights}{Logical; whether to change the size of the edges indicating connections}

\item{minNFlights}{Numeric; plot connections with at least this specified number of flights}

\item{map}{String or data.frame or NULL as a background map}

\item{vertexColors}{Vector with IATA codes used as colors for the vertices/airports}

\item{vertexShapes}{Vector with IATA codes used as shapes for the vertices/airports}

\item{xyRatio}{Numeric; approximate X-Y ratio of the area}

\item{clipMap}{Logical or numeric scalar; whether to ignore the map image when specifying the axis limits of the figure}

\item{useLatex}{Logical; whether to format numbers as latex code}

\item{edgeAlpha}{0 or 1; the alpha value used in plotting edges/connections}
}
\value{
A \code{ggp} object
}
\description{
\code{plotFlights_adj()} creates useful figures such as the flight connections from the flight delay dataset.
This function is a slight adjusted version of the function \code{\link[=plotFlights]{plotFlights()}} in \code{graphicalExtremes} package.
}
\references{
Engelke S, Hitz A, Gnecco N, Hentschel M (2023). \emph{graphicalExtremes:
Statistical Methodology for Graphical Extreme Value Models}. R package
version 0.3.0, \url{https://github.com/sebastian-engelke/graphicalExtremes}
}
