% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoCop.R
\name{PseudoCop}
\alias{PseudoCop}
\title{Redefining the conditional distribution function of pair-copulas}
\usage{
PseudoCop(j, cind, data, StrMtx, fam, MaxMtr, Par)
}
\arguments{
\item{j}{Integer; the number of conditioning variables in the conditioning set \eqn{D(e)}}

\item{cind}{Integer; stands for the column index corresponding to the \eqn{i}th column.
Given the \eqn{i}th column, subset the data given all variables in \eqn{D(e)} are less than 1.}

\item{data}{An \eqn{N\times d} data matrix of either pseudo-observations or random samples from the inverted-Pareto distribution}

\item{StrMtx}{A \eqn{d \times d} specified structure matrix}

\item{fam}{A \eqn{d \times d} specified family matrix}

\item{MaxMtr}{A \eqn{d\times d} max-matrix}

\item{Par}{A \eqn{d\times d} recursively estimated parameter matrix from previous trees}
}
\value{
A list of two numeric vectors, including two conditional distribution functions of pair-copulas:
\itemize{
\item VdirC
\item VindirC
}
}
\description{
\code{PseudoCop} recalculates the conditional distribution function of pair-copulas by conditioning on all variables
in the conditioning set being less than 1.
This calculation is based on recursive approaches tree by tree sequentially from \eqn{T_3} to \eqn{T_{d-1}}.
For more details, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.
}
