% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderStrMtx.R
\name{ReorderStrMtx}
\alias{ReorderStrMtx}
\title{Permute the diagonal elements of a structure matrix}
\usage{
ReorderStrMtx(Matrix, oldOrder = NULL)
}
\arguments{
\item{Matrix}{Numeric \eqn{d \times d} structure matrix}

\item{oldOrder}{Logical; If \code{NULL} (the default), the function uses the order
of diagonal elements in the structure matrix.}
}
\value{
Numeric \eqn{d \times d} structure matrix with diagonal elements
arranged in ascending order.
}
\description{
\code{ReorderStrMtx()} reorders the indices of variables to ensure that the
diagonal elements of a structure matrix are arranged in ascending order.
}
\examples{
##  A 5 x 5 structure matrix where diagonal elements are not arranged in ascending order.
StrMtx <- cbind(c(2,0,0,0,0),c(2,4,0,0,0),c(4,2,5,0,0),c(2,4,5,3,0),c(2,3,4,5,1))

##  A 5 x 5 structure matrix with diagonal elements arranged in increasing order.
ReorderStrMtx(StrMtx)
}
