% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revert.R
\name{revert}
\alias{revert}
\title{Convert a lower (or upper) vine structure matrix}
\usage{
revert(m)
}
\arguments{
\item{m}{A numeric \eqn{d\times d} lower (or upper) triangular matrix specifying
the regular vine structure.}
}
\value{
A numeric \eqn{d\times d} upper triangular matrix if the input matrix is lower
triangular; otherwise, it is a lower triangular matrix.
}
\description{
\code{revert()} transforms a lower (or upper) triangular vine structure matrix
to the corresponding upper (or lower) triangular matrix.
}
\examples{
##  A 5 x 5 lower triangular structure matrix
LowerStrMtx <- matrix(c(5, 2, 3, 1, 4,
 0, 2, 3, 4, 1,
 0, 0, 3, 4, 1,
 0, 0, 0, 4, 1,
 0, 0, 0, 0, 1),5,5)
 
##  Convert to the upper triangular matrix
revert(LowerStrMtx)
}
