% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardStrMtx.R
\name{standardStrMtx}
\alias{standardStrMtx}
\title{Reconstruct a structure matrix with diagonal elements put in ascending order}
\usage{
standardStrMtx(Matrix, oldOrder = NULL)
}
\arguments{
\item{Matrix}{A numeric \eqn{d \times d} structure matrix}

\item{oldOrder}{Logical; If \code{NULL} (the default), the function uses the order
of diagonal elements in the structure matrix.}
}
\value{
A numeric \eqn{d \times d} structure matrix with diagonal elements
arranged in ascending order.
}
\description{
\code{standardStrMtx()} reconstructs a given structure matrix to ensure that the
diagonal elements of a structure matrix are arranged in ascending order.
The advantage of this reconstruction is that we can create a so-called max-matrix (see:\code{createMaxMtx()})
, making it easier for the algorithm to find the correct argument for conditional pair-copulas when recursive relationships are used.
}
\examples{
##  A 5 x 5 structure matrix where diagonal elements are not arranged in ascending order.
StrMtx <- cbind(c(2,0,0,0,0),c(2,4,0,0,0),c(4,2,5,0,0),c(2,4,5,3,0),c(2,3,4,5,1))

##  A 5 x 5 structure matrix with diagonal elements arranged in increasing order.
standardStrMtx(StrMtx)
}
