% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeVarInd.R
\name{storeVarInd}
\alias{storeVarInd}
\title{Store the positions of variable indices in a structure matrix}
\usage{
storeVarInd(Diag)
}
\arguments{
\item{Diag}{Numeric \eqn{d \times d} structure matrix.}
}
\value{
A numeric vector containing the locations of variable indices.
}
\description{
\code{storeVarInd()} stores the positions of variable indices (\eqn{1,\ldots,d})
in a structure matrix.
The output vector will be used to rearrange the order of generated variables
from \code{\link[=XVineSim]{XVineSim()}}, which may not necessarily align with \eqn{1,\ldots,d}.
}
\examples{
##  
}
