% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcFamSel.R
\name{tcFamSel}
\alias{tcFamSel}
\title{Select the family of bivariate tail copula models for each edge and fit the model}
\usage{
tcFamSel(x1, x2, si = 0.9, familyset, selectioncrit = "AIC")
}
\arguments{
\item{x1}{A numeric vector; indicates the first variable corresponding to the first node of the edge \eqn{e=\{a,b\}} in the first tree.}

\item{x2}{A numeric vector; indicates the second variable corresponding to the second node of the edge \eqn{e=\{a,b\}} in the first tree.}

\item{si}{Numeric; a tuning parameter for mBIC (\eqn{\si_0=0.9}; default).}

\item{familyset}{Numeric vector; indicates the list of bivariate tail copula models.}

\item{selectioncrit}{Character string; indicates the selection criterion for the family of bivariate tail copula models.}
}
\value{
A list object consisting of the class type, parameter estimates, log-likelihood value, AIC, pseudo-observations
for selected bivariate tail copula models.
}
\description{
Given the family of bivariate tail copula models users consider, \code{tcFamSel()} selects
the family of bivariate tail copula models with the lowest AIC for a specific edge
, and returns parameter estimates, likelihood value, AIC, and pseud-observations used for the next tree.
}
