% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeCrit.R
\name{TreeCrit}
\alias{TreeCrit}
\title{Specify the selection criterion for regular vine trees}
\usage{
TreeCrit(treecrit, famset)
}
\arguments{
\item{treecrit}{A function that includes three arguments (u1,u2,weights). The first two arguments are for bivariate parametric models and the last argument 'weights' can be
used for weighting missing values.}

\item{famset}{A numeric vector; the list of families for pair-copula models.}
}
\value{
The defined function \code{treecrit} used to calculate edge weights.
}
\description{
\code{TreeCrit()} chooses the selection criterion for regular vine trees. Possible criterion for the first tree
are the lower tail dependence measure called \code{chi} and the \code{variogram} for the Husler-Reiss distribution.
Other criteria for subsequent trees are Kendall's tau, Spearsman rho, information critera such as AIC, BIC, and adjusted AIC.
This function is built upon the function \code{VineCopula:::set_treecrit} in \code{VineCopula} package.
Users can add additional tree criteria to the function.
}
\examples{
TreeCrit('chi')
}
