% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xvine-package.R
\docType{package}
\name{Xvine-package}
\alias{Xvine}
\alias{Xvine-package}
\title{Xvine: Statistical Modelling for Multivariate Extremes via Vine Structures}
\description{
Modeling multivariate extremes via vine structures provides flexibility and sparse structures in high dimensions. Using recursive approaches, we perform parameter estimation, select bivariate (tail) copulas, and determine the sequence of the vine tree.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeongjinLee88/Xvine}
  \item \url{https://jeongjinlee88.github.io/Xvine/}
  \item Report bugs at \url{https://github.com/JeongjinLee88/Xvine/issues}
}

}
\author{
\strong{Maintainer}: Jeongjin Lee \email{stecophil88@gmail.com} (\href{https://orcid.org/0000-0001-5655-3358}{ORCID})

Authors:
\itemize{
  \item Anna Kiriliouk \email{anna.kiriliouk@unamur.be} (\href{https://orcid.org/0000-0003-2977-0140}{ORCID})
  \item Johan Segers \email{johan.segers@uclouvain.be} (\href{https://orcid.org/0000-0002-0444-689X}{ORCID})
}

}
\keyword{internal}
