% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVineBoxPlot.R
\name{XVineBoxplot}
\alias{XVineBoxplot}
\title{Box-plots of parameter estimates for X-vine models}
\usage{
XVineBoxplot(
  N,
  qt,
  ite,
  XVS,
  RankT = TRUE,
  familyset_tc = c(1:4),
  familyset_cop = c(0, 1, 3, 4, 5, 6, 13, 14, 16)
)
}
\arguments{
\item{N}{Numeric; the sample size.}

\item{qt}{Numeric; a lower quantile as a threshold.}

\item{ite}{Numeric; the number of iterations for generating samples.}

\item{XVS}{A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:\code{\link[=XVineSpec]{XVineSpec()}}.}

\item{RankT}{Logical; select either rank-based data (default) or data directly from the limiting distribution.}

\item{familyset_tc}{Numeric vector; the model type of bivariate tail copulas.
Possible tail copula models are:
\itemize{
\item 1=Husler-Reiss model
\item 2=Negative logistic model
\item 3=Logistic model
\item 4=Dirichlet model
}}

\item{familyset_cop}{Numeric vector; the class of bivariate copula families with a single parameter}
}
\value{
A list of parameter estimates matrices for repeated simulations and boxplots of the parameter estimates
where the first \eqn{d-1} plots are for the first level tree \eqn{T_1} and the next d-2 plots are for \eqn{T_2} and so on.
\code{bp_mle_SPE} is the boxplot of ML estimates whereas \code{bp_mle_Fam} plots the boxplot of ML estimates by selecting the family of bivariate parametric models.
Similarly, \code{bp_dep_SPE} gives the boxplot of dependence measures and \code{bp_dep_Fam} shows the boxplot of dependence measures by selecting the family of bivariate parametric models.
}
\description{
\code{XVineBoxplot()} creates box-plots of parameter estimates for X-Vine models.
We implement the maximum likelihood method for estimating parameters.
The function provides ML estimates as well as dependence measure coefficients converted from the ML estimates.
Given the regular vine structure, the function considers estimating parameters by either selecting the family of bivariate (tail) copula via the AIC
or using the specified family.
Four possible tail copula families are available:
Husler-Reiss, Negative logistic, logistic, and Dirichlet models.
For explicit formulas for the (lower) tail dependence coefficient \eqn{\chi_{L,e}=R(1,1)} for \eqn{e\in T_1} from the ML estimate, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023).
The function converts the ML estimate to the measure of Kendall's tau, \eqn{\tau_{e}}, for \eqn{e\in T_i}, \eqn{i=2,\ldots,d-1}.
Any copula family with a single parameter can be considered. For more details, refer to Czado (2019, Section 3.5).
}
\examples{
# Specify the number of iterations
ite=200
# Specify the structure matrix
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
# Specify the parameter matrix
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
# Specify the family matrix                 
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# To run, uncomment it.
# XVineBoxplot(N = 500, qt = 0.05, ite = ite, XVS = XVS, RankT = TRUE)
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.
Czado, C. (2019). Analyzing dependent data with vine copulas. Lecture Notes in Statistics, Springer, 222.
}
