% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVineFamSelSimul.R
\name{XVineFamSelSimul}
\alias{XVineFamSelSimul}
\title{Simulation for the selection of bivariate parametric families}
\usage{
XVineFamSelSimul(
  N,
  Rank = TRUE,
  qt,
  ite,
  XVS,
  edgeIndex,
  familyset_tc,
  familyset_cop,
  selectioncrit = "AIC",
  effsampsize = 10,
  tau_threshold = 0.05
)
}
\arguments{
\item{N}{Numeric; the sample size.}

\item{Rank}{Logical; whether rank transformation is performed or not (\code{Rank=T}; default).}

\item{qt}{Numeric; a lower enough threshold used in the rank transformation (e.g. switches from Pareto scale to uniform scale)}

\item{ite}{Numeric; the number of iterations for simulation.}

\item{XVS}{A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:\code{\link[=XVineSpec]{XVineSpec()}}.}

\item{edgeIndex}{Numeric vector; the index set that indicates the location of the matrix for the corresponding edge.}

\item{familyset_tc}{Numeric vector; the model type of bivariate tail copulas.
Possible tail copula models are:
\itemize{
\item 1=Husler-Reiss model
\item 2=Negative logistic model
\item 3=Logistic model
\item 4=Dirichlet model
}}

\item{familyset_cop}{Numeric vector; the class of bivariate copula families with a single parameter.}

\item{selectioncrit}{Character; the information criteria for model selections (default: \code{AIC}).}

\item{effsampsize}{Integer; indicates the effective sample size (Default: 10) for the independence copula.}

\item{tau_threshold}{Numeric; indicates the value of the Kendall's tau (Default: 0.05) for the independence copula.}
}
\value{
\enumerate{
\item \code{FamSelOut}: A list of \eqn{d\times d} strict upper triangular matrices from the function \code{\link[=XVineFamSel]{XVineFamSel()}} for iterations.
\item \code{EffSizePercent}: A \eqn{d\times d} strict upper triangular matrix including the percentages of the effective sample size.
\item \code{GlobalProp}: Numeric; the proportion of correctly selected families over all trees.
\item \code{TreeProp}: Numeric vector; the proportion of correctly selected families for each level tree.
\item \code{EdgeProp}: A \eqn{d\times d} strict upper triangular matrix including the proportion of correctly selected families for each edge in each tree.
\item \code{SelectIndCop}: Numeric; the number of times the algorithm chooses the independence copula for the specified edge by \code{edgeIndex}.
}
}
\description{
\code{XVineFamSelSimul} returns simulation results of selecting bivariate parametric families
by calling the function \code{\link[=XVineFamSel]{XVineFamSel()}}.
All information about the simulation result is saved in the list of \code{FamSelOut}.
}
\examples{
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# Approx 50 secs for N=4000
XVineFamSelSimul(N = 1000, qt = 0.05, ite = 100, XVS = XVS
,edgeIndex = c(4,5), familyset_tc=c(1:4), familyset_cop = c(0,1,3,4,5,6,13,14,16))
}
