% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVineSim.R
\name{XVineSim}
\alias{XVineSim}
\title{Simulation from X-Vine models}
\usage{
XVineSim(N, XVS, k)
}
\arguments{
\item{N}{Integer; the number of samples to simulate.}

\item{XVS}{A list of three matrix components for each conditioning variable:
\itemize{
\item reproduced structure matrices
\item family matrices
\item parameter matrices.
To specify the argument \code{XVS}, see \code{\link[=XVineSpec]{XVineSpec()}}.
}}

\item{k}{Integer; the \eqn{k}th, \eqn{k\in {1,\ldots,d}}, conditioning index to condition on.}
}
\value{
An \eqn{N\times d} data matrix used to calculate extremal functions in \code{\link[=ParetoSim]{ParetoSim()}}.
}
\description{
\code{XVineSim()} generates multivariate samples from X-Vine models
given one of the variables is less than 1.
The simulation algorithm selects the X-vine specification
associated with the conditioned variable and uses
recursive approaches to return an \eqn{N \times d} data matrix.
For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)
}
\examples{
##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.
}
