% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVineSpec.R
\name{XVineSpec}
\alias{XVineSpec}
\title{Reproduce the given X-vine matrix specification depending on each conditioning variable}
\usage{
XVineSpec(M, Mmod, Mpar)
}
\arguments{
\item{M}{A \eqn{d \times d} upper triangular structure matrix.}

\item{Mmod}{A \eqn{d \times d} strict upper triangular family matrix for identifying
bivariate (tail) copulas.}

\item{Mpar}{A \eqn{d \times d} strict upper triangular parameter matrix associated with the family matrix.}
}
\value{
A list of arrays includes:
\itemize{
\item xmat: array of dimension \eqn{d \times d \times d} containing rearranged X-vine structure matrices.
\item fmat: array of dimension \eqn{d \times d \times d} containing family matrices associated with the structure matrices.
\item pmat: array of dimension \eqn{d \times d \times d} containing parameter matrices associated with the structure matrices.
\item xedge: A list of edges corresponding to the initial structure matrix.
}
}
\description{
Given the initial X-vine matrix specification of (structure matrix, family matrix, parameter matrix),
\code{XVineSpec()} generates a list of redefined X-vine specifications for each conditioning variable.
Preserving the vine structure, the function reconstructs structure matrices where the first diagonal element
, denoted as \eqn{m_{11}\in [d]}, corresponds to the conditioning variable.
Additionally, \code{XVineSpec()} includes the associated parameter and family matrices based on the given structure matrices.
}
\details{
To use the exact simulation algorithm for multivariate Pareto distribution in Engelke and Hitz (2020) and Kiriliouk et al. (2023),
we condition on each variable to calculate extremal functions.
For each conditioning variable, the index of the conditioning variable is placed on the first diagonal element in the structure matrix.
Specifically, in a \eqn{d \times d} structure matrix, the element \eqn{m_{jj}\in [d]} corresponds to the index of the conditioning variable \eqn{X_j} for \eqn{j=1,\ldots,d}.
All reproduced structure matrices maintain the same vine tree structure.
The corresponding parameter matrices contain the parameter values for the bivariate (tail) copulas.
In family matrices, the first row specifies bivariate tail copulas and the remaining rows specify bivariate copulas.
}
\examples{
##  A 6 x 6 structure matrix
M <- cbind(c(1,0,0,0,0,0), c(1,4,0,0,0,0), c(4,1,6,0,0,0),
          c(4,1,6,3,0,0), c(6,4,1,3,5,0), c(3,4,1,6,5,2))
##  A 6 x 6 family matrix
Mmod <- cbind(rep(0,6),c("neglog", rep(0,5)),
              c("hr", "clayton",  rep(0,4)),
              c("log", "gumbel", "joe", rep(0,3)),
              c("dir", "frank", "clayton", "gumbel", 0, 0),
              c("neglog", "joe", "frank", "clayton", "gumbel", 0))
##  A 6 x 6 parameter matrix
Mpar <- cbind(rep(0,6),c(0.25, rep(0,5)), c(0.3, 0.7,  rep(0,4)),
              c(0.5,0.7,0.1, rep(0,3)), c(0.2,0.8,0.3,0.55, 0, 0),
              c(0.1,0.9,0.75,0.3,0.45, 0))
##  Permute the given X-vine specification
PermXVineSpec <- XVineSpec(M,Mmod,Mpar)
PermXVineSpec[[1]]
cbind(PermXVineSpec[[1]][,,5],PermXVineSpec[[2]][,,5])
##  Compare with cbind(M,Mmod): looks ok!
##  Truncated X-vine specification at tree level 3
M <- cbind(c(1,0,0,0,0,0), c(1,4,0,0,0,0), c(4,1,6,0,0,0),
           c(4,1,6,3,0,0), c(6,4,1,0,5,0), c(3,4,1,0,0,2))
Mmod <- cbind(rep(0,6),c("neglog", rep(0,5)), c("hr", "clayton",  rep(0,4)), 
           c("log", "gumbel", "joe", rep(0,3)), c("dir", "frank", "clayton", 0, 0, 0),
           c("neglog", "joe", "frank", 0,0, 0))
Mpar <- cbind(rep(0,6),c(0.25, rep(0,5)), c(0.3, 0.7,  rep(0,4)), 
              c(0.5,0.7,0.1, rep(0,3)), c(0.2,0.8,0.3,0, 0, 0),
              c(0.1,0.9,0.75,0,0, 0))
PermTrucXVine <- XVineSpec(M,Mmod,Mpar)
PermTrucXVine[[1]]
cbind(PermTrucXVine[[1]][,,5],PermTrucXVine[[2]][,,5]) #compare with cbind(M,Mmod): looks ok!
}
\references{
Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Engelke, S., & Hitz, A. S. (2020). Graphical models for extremes. Journal of the Royal Statistical Society Series B: Statistical Methodology, 82(4), 871-932.
}
\seealso{
\code{XVineEdges()}
}
