% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMaxMtx.R
\name{createMaxMtx}
\alias{createMaxMtx}
\title{Create a maximum matrix from a structure matrix with diagonal elements in ascending order}
\usage{
createMaxMtx(Matrix)
}
\arguments{
\item{Matrix}{Numeric \eqn{d \times d} upper triangular matrix that specifies
the regular vine structure.}
}
\value{
Numeric \eqn{d \times d} upper triangular max-matrix.
}
\description{
\code{createMaxMtx()} creates a max-matrix from a structure matrix where diagonal elements are put in increasing order.
This max-matrix is used in sequential parameter estimation
to determine the appropriate argument for conditional pair-copulas.
For more details, refer to Chapter 6 in Czado (2019).
}
\examples{
##  Create a 5 x 5 upper triangular structure matrix
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
##  Derive the corresponding 5 x 5 upper triangular max-matrix
MaxMtx <- createMaxMtx(StrMtx)
}
\references{
Czado, C. (2019). Analyzing dependent data with vine copulas.
Lecture Notes in Statistics, Springer, 222.
}
